requirements.txt
flask==2.3.2
python-docx==1.1.2
gunicorn==21.2.0
